% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{conditional}
\alias{conditional}
\title{test if `::` on packages in Suggests causes R CMD check warning}
\usage{
conditional(...)
}
\arguments{
\item{...}{arguments, you know}
}
\description{
test if `::` on packages in Suggests causes R CMD check warning
}
